Separate stores in a different folder
